refactor(api) migrate user existence verification to src
