+++ api/src/identity-access-management/application/user/user-existence-verification-pre-handler.js
@@ -1,6 +1,6 @@
-import * as userRepository from '../../../src/identity-access-management/infrastructure/repositories/user.repository.js';
-import * as errorSerializer from '../../../src/shared/infrastructure/serializers/jsonapi/validation-error-serializer.js';
-import { UserNotFoundError } from '../../domain/errors.js';
+import { UserNotFoundError } from '../../../../lib/domain/errors.js';
+import * as errorSerializer from '../../../shared/infrastructure/serializers/jsonapi/validation-error-serializer.js';
+import * as userRepository from '../../infrastructure/repositories/user.repository.js';
 
 const verifyById = function (request, h, dependencies = { userRepository, errorSerializer }) {
   return dependencies.userRepository.get(request.params.id).catch((err) => {
