+++ api/src/identity-access-management/application/user/user.route.js
@@ -1,13 +1,13 @@
 import Joi from 'joi';
 import XRegExp from 'xregexp';
 
-import { userVerification } from '../../../../lib/application/preHandlers/user-existence-verification.js';
 import { securityPreHandlers } from '../../../shared/application/security-pre-handlers.js';
 import { config } from '../../../shared/config.js';
 import { EntityValidationError } from '../../../shared/domain/errors.js';
 import { AVAILABLE_LANGUAGES } from '../../../shared/domain/services/language-service.js';
 import { identifiersType } from '../../../shared/domain/types/identifiers-type.js';
 import { userController } from './user.controller.js';
+import { userVerification } from './user-existence-verification-pre-handler.js';
 
 const { passwordValidationPattern } = config.account;
