+++ api/tests/identity-access-management/unit/application/user-existence-verification_test.js
@@ -1,5 +1,5 @@
-import { userVerification } from '../../../../lib/application/preHandlers/user-existence-verification.js';
 import { UserNotFoundError } from '../../../../lib/domain/errors.js';
+import { userVerification } from '../../../../src/identity-access-management/application/user/user-existence-verification-pre-handler.js';
 import { expect, hFake, sinon } from '../../../test-helper.js';
 
 describe('Unit | Pre-handler | User Verification', function () {
